refactor reverse_shuffle_table
