+++ bodo/hiframes/pd_groupby_ext.py
@@ -41,7 +41,6 @@ from bodo.libs.array import (
     delete_table,
     get_null_shuffle_info,
     get_shuffle_info,
-    reverse_shuffle_table,
     shuffle_table,
 )
 from bodo.libs.array_item_arr_ext import ArrayItemArrayType
@@ -2555,6 +2554,29 @@ def gen_shuffle_dataframe(df, keys, _is_parallel):
 
 def reverse_shuffle(data, shuffle_info):  # pragma: no cover
     return data
+
+
+@intrinsic
+def reverse_shuffle_table(typingctx, table_t, shuffle_info_t):
+    """call reverse shuffle if shuffle info not none"""
+    from llvmlite import ir as lir
+
+    from bodo.libs.array import table_type
+
+    def codegen(context, builder, sig, args):
+        if sig.args[-1] == types.none:
+            return context.get_constant_null(table_type)
+
+        fnty = lir.FunctionType(
+            lir.IntType(8).as_pointer(),
+            [lir.IntType(8).as_pointer(), lir.IntType(8).as_pointer()],
+        )
+        fn_tp = cgutils.get_or_insert_function(
+            builder.module, fnty, name="reverse_shuffle_table"
+        )
+        return builder.call(fn_tp, args)
+
+    return table_type(table_type, shuffle_info_t), codegen
 
 
 @overload(reverse_shuffle)
